Revert "Merge branch 'main' into Adam's-Branch"
This reverts commit 975c0844f60accd20acdd7b6fdab19173989ec19, reversing
changes made to e3946f50ba824c0326ecfa4158227deb7a01bcda.

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/AccountRoleRepository.java
@@ -5,5 +5,5 @@ import ca.mcgill.ecse321.grocerystore.model.AccountRole;
 
 public interface AccountRoleRepository extends CrudRepository<AccountRole, Integer> {
 	
-	AccountRole findByRoleID(Long roleID);
+	AccountRole findByRoleID(Integer roleID);
 }
