+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/CartRepository.java
@@ -7,7 +7,7 @@ import ca.mcgill.ecse321.grocerystore.model.Cart;
 
 public interface CartRepository extends CrudRepository<Cart, Integer> {
 
-	Cart findByCartID(Long cartID);
+	Cart findByCartID(Integer cartID);
 	
 	Cart findByAccount(Account account);
 }
