+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/ClerkRepository.java
@@ -7,5 +7,5 @@ import ca.mcgill.ecse321.grocerystore.model.Clerk;
 
 public interface ClerkRepository extends CrudRepository<Clerk, Integer> {
 	
-	Clerk findByRoleID(Long roleID);
+	Clerk findByRoleID(Integer roleID);
 }
