+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/EmployeeRepository.java
@@ -6,5 +6,5 @@ import ca.mcgill.ecse321.grocerystore.model.Employee;
 
 public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
 	
-	Employee findByRoleID(Long roleID);
+	Employee findByRoleID(Integer roleID);
 }
