+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/OrderRepository.java
@@ -6,5 +6,5 @@ import ca.mcgill.ecse321.grocerystore.model.Order;
 
 public interface OrderRepository extends CrudRepository<Order, Integer> {
 
-	Order findByOrderID(Long orderID);
+	Order findByOrderID(Integer orderID);
 }
