+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/StoreRepository.java
@@ -5,5 +5,6 @@ import ca.mcgill.ecse321.grocerystore.model.Store;
 
 public interface StoreRepository extends CrudRepository<Store,Integer>{
 
-		Store findByStoreID(Long storeID);	
+		Store findByStoreID(Integer storeID);
+		
 }
