+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/TerminalRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import ca.mcgill.ecse321.grocerystore.model.Terminal;
 
 public interface TerminalRepository extends CrudRepository<Terminal,Integer>{
-	
-	Terminal findByTerminalID(Long terminalID);
+	Terminal findByTerminalID(Integer terminalID);
 }
