+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/TimeSlotRepository.java
@@ -6,5 +6,5 @@ import ca.mcgill.ecse321.grocerystore.model.TimeSlot;
 
 public interface TimeSlotRepository extends CrudRepository<TimeSlot,Integer>{
 
-	TimeSlot findByTimeSlotID(Long timeSlotID);
+	TimeSlot findByTimeSlotID(Integer timeSlotID);
 }
