+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dto/TerminalDto.java
@@ -2,17 +2,17 @@ package ca.mcgill.ecse321.grocerystore.dto;
 
 public class TerminalDto {
     // attribute
-    private Long terminalID;
+    private Integer terminalID;
 
     // constructor
     public TerminalDto() {
     }
 
-    public TerminalDto(Long terminalID){
+    public TerminalDto(Integer terminalID){
         this.terminalID = terminalID;
     }
     // getter
-    public Long getTerminalID() {
+    public Integer getTerminalID() {
         return terminalID;
     }
 }
